slamd-96: Started moving preprocessing functions into ExperimentPreprocessor class
